apps/ml-service/main.py
"""
ML Service for Stock Price Prediction using LSTM
"""
from fastapi import FastAPI, HTTPException
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from typing import List, Dict, Optional
import numpy as np
from datetime import datetime
import os
from dotenv import load_dotenv

load_dotenv()

app = FastAPI(
    title="Stock Prediction ML Service",
    description="LSTM-based stock price prediction API",
    version="1.0.0"
)

# CORS middleware
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Request/Response Models
class PredictionRequest(BaseModel):
    symbol: str
    timeframes: List[str] = ["1h", "4h", "1d", "1w"]

class PredictionResponse(BaseModel):
    symbol: str
    current_price: float
    timestamp: int
    predictions: List[Dict]
    indicators: Dict
    confidence: float
    ai_analysis: Dict

class ModelInfoResponse(BaseModel):
    model_version: str
    last_trained: str
    accuracy: Dict
    supported_symbols: List[str]

# Global model instance (will be loaded on startup)
model_instance = None

@app.on_event("startup")
async def startup_event():
    """Load ML models on startup"""
    global model_instance
    # TODO: Load actual trained models
    print(" Loading ML models...")
    model_instance = {"status": "mock"}  # Placeholder
    print(" ML models loaded")

@app.get("/")
async def root():
    """Health check endpoint"""
    return {
        "status": "ok",
        "service": "ML Prediction Service",
        "version": "1.0.0",
        "timestamp": datetime.now().isoformat()
    }

@app.get("/health")
async def health():
    """Detailed health check"""
    return {
        "status": "healthy",
        "model_loaded": model_instance is not None,
        "timestamp": datetime.now().isoformat()
    }

@app.post("/predict", response_model=PredictionResponse)
async def predict(request: PredictionRequest):
    """Generate price predictions for a stock symbol"""
    try:
        symbol = request.symbol.upper()
        
        # TODO: Implement actual LSTM prediction
        # For now, return mock predictions
        current_price = 178.32
        
        predictions = []
        timeframe_configs = {
            "1h": {"hours": 1, "confidence": 87},
            "4h": {"hours": 4, "confidence": 79},
            "1d": {"hours": 24, "confidence": 72},
            "1w": {"hours": 168, "confidence": 65}
        }
        
        for tf in request.timeframes:
            if tf not in timeframe_configs:
                continue
                
            config = timeframe_configs[tf]
            # Mock prediction logic
            predicted_price = current_price * (1 + np.random.uniform(-0.03, 0.05))
            change = predicted_price - current_price
            change_percent = (change / current_price) * 100
            
            predictions.append({
                "timeframe": tf,
                "predicted_price": round(predicted_price, 2),
                "confidence": config["confidence"],
                "change": round(change, 2),
                "change_percent": round(change_percent, 2),
                "upper_bound": round(predicted_price * 1.02, 2),
                "lower_bound": round(predicted_price * 0.98, 2)
            })
        
        # Mock technical indicators
        indicators = {
            "rsi": {"value": 58.3, "status": "neutral"},
            "macd": {"value": 2.45, "status": "bullish"},
            "bollinger": {"value": 0.78, "status": "bullish"},
            "volume": {"value": 125.5, "status": "bullish"}
        }
        
        # AI analysis
        ai_analysis = {
            "sentiment": "bullish",
            "confidence": 87,
            "reasoning": "Strong bullish momentum detected with positive MACD crossover and RSI in neutral territory. Volume is increasing, indicating strong buyer interest.",
            "risk_factors": [
                "High market volatility expected",
                "Earnings report in 2 days",
                "Technical resistance at $180"
            ]
        }
        
        return PredictionResponse(
            symbol=symbol,
            current_price=current_price,
            timestamp=int(datetime.now().timestamp() * 1000),
            predictions=predictions,
            indicators=indicators,
            confidence=87.0,
            ai_analysis=ai_analysis
        )
        
    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.get("/model/info", response_model=ModelInfoResponse)
async def model_info():
    """Get information about the ML model"""
    return ModelInfoResponse(
        model_version="1.0.0-lstm",
        last_trained="2025-10-14T00:00:00Z",
        accuracy={
            "1h": 92.3,
            "4h": 88.7,
            "1d": 85.2,
            "1w": 78.9,
            "overall": 87.5
        },
        supported_symbols=["AAPL", "GOOGL", "MSFT", "TSLA", "AMZN", "NVDA", "META", "NFLX"]
    )

@app.post("/model/train")
async def train_model(symbol: str):
    """Trigger model training for a specific symbol"""
    # TODO: Implement actual model training
    return {
        "status": "training_started",
        "symbol": symbol.upper(),
        "estimated_time": "30 minutes",
        "message": "Model training initiated in background"
    }

@app.get("/indicators/{symbol}")
async def get_indicators(symbol: str):
    """Calculate technical indicators for a symbol"""
    # TODO: Implement actual indicator calculations
    return {
        "symbol": symbol.upper(),
        "timestamp": int(datetime.now().timestamp() * 1000),
        "indicators": {
            "rsi": 58.3,
            "macd": 2.45,
            "macd_signal": 1.22,
            "macd_histogram": 1.23,
            "bollinger_upper": 185.5,
            "bollinger_middle": 178.3,
            "bollinger_lower": 171.1,
            "sma_20": 176.5,
            "sma_50": 172.8,
            "sma_200": 165.3,
            "ema_12": 177.2,
            "ema_26": 174.5,
            "atr": 3.45,
            "stochastic_k": 68.5,
            "stochastic_d": 65.3
        }
    }

if __name__ == "__main__":
    import uvicorn
    port = int(os.getenv("PORT", 8001))
    host = os.getenv("HOST", "0.0.0.0")
    debug = os.getenv("DEBUG", "False").lower() == "true"
    
    uvicorn.run(
        "main:app",
        host=host,
        port=port,
        reload=debug,
        log_level="info"
    )