apps/ml-service/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
numpy==1.26.4
pandas==2.2.1
scikit-learn==1.4.1
tensorflow==2.15.0
ta==0.11.0
python-dotenv==1.0.1
httpx==0.27.0
redis==5.0.2